crd: Refactor RegisterCRDsCell to be extensible
Current implementation of RegisterCRDsCell is strictly tied to the
registration of the CRDs belonging to the "cilium.io" group.

This commit refactors the cell to make it easily extendable if other
groups needs to register their CRDs at operator startup.
